refactor(ai-agent): extract event_tracker from acp/agent.rs
