+++ crates/aionui-ai-agent/src/manager/acp/agent.rs
@@ -71,7 +71,7 @@ pub struct AcpAgentManager {
     /// Single in-memory source of truth for session lifecycle, modes,
     /// models, config, and all runtime data previously split across
     /// `AcpRuntimeSnapshot` and `AcpState`.
-    session: RwLock<AcpSession>,
+    pub(super) session: RwLock<AcpSession>,
     /// Standalone conversation status (not part of the session aggregate
     /// because it is a UI-level concern, not ACP protocol state).
     status: RwLock<Option<ConversationStatus>>,
@@ -80,7 +80,7 @@ pub struct AcpAgentManager {
     /// ACP protocol handle (SDK connection).
     protocol: AcpProtocol,
     /// Typed event broadcast channel.
-    event_tx: broadcast::Sender<AgentStreamEvent>,
+    pub(super) event_tx: broadcast::Sender<AgentStreamEvent>,
     /// Timestamp of last activity (atomic for lock-free reads). Shared
     /// with the `PermissionRouter` so permission arrivals update the
     /// activity timestamp without reverse-referencing the manager.
@@ -94,7 +94,7 @@ pub struct AcpAgentManager {
     skill_manager: Arc<AcpSkillManager>,
     /// Domain event sender — session aggregate events are forwarded here
     /// for the persistence consumer (`AcpSessionSyncService`).
-    domain_event_tx: mpsc::Sender<AcpSessionEvent>,
+    pub(super) domain_event_tx: mpsc::Sender<AcpSessionEvent>,
 }
 
 impl AcpAgentManager {
@@ -338,65 +338,9 @@ impl AcpAgentManager {
             .start(self.event_tx.clone(), Arc::clone(&self.last_activity));
     }
 
-    /// Start the session event tracker loop.
-    pub fn start_session_event_tracker(self: &Arc<Self>) {
-        let this = Arc::clone(self);
-        tokio::spawn(async move {
-            let mut rx = this.event_tx.subscribe();
-            loop {
-                match rx.recv().await {
-                    Ok(event) => {
-                        this.apply_event_to_session(&event).await;
-                    }
-                    Err(broadcast::error::RecvError::Closed) => break,
-                    Err(broadcast::error::RecvError::Lagged(_)) => continue,
-                }
-            }
-        });
-    }
-
-    /// Mirror a stream event into the `AcpSession` aggregate's observed/advertised
-    /// layer and forward any resulting domain events to the persistence consumer.
-    async fn apply_event_to_session(&self, event: &AgentStreamEvent) {
-        match event {
-            AgentStreamEvent::AcpModeInfo(value) => {
-                if let Ok(update) = serde_json::from_value::<SessionModeState>(value.clone()) {
-                    let mut s = self.session.write().await;
-                    s.apply_advertised_modes(update);
-                    self.commit_session_changes(&mut s).await;
-                } else if let Some(current_id) = value.get("currentModeId").and_then(|v: &Value| v.as_str()) {
-                    let mut s = self.session.write().await;
-                    s.apply_observed_mode(ModeId::new(current_id));
-                    self.commit_session_changes(&mut s).await;
-                }
-            }
-            AgentStreamEvent::AcpModelInfo(value) => {
-                if let Ok(update) = serde_json::from_value::<SessionModelState>(value.clone()) {
-                    let mut s = self.session.write().await;
-                    s.apply_advertised_models(update);
-                    self.commit_session_changes(&mut s).await;
-                }
-            }
-            AgentStreamEvent::AcpConfigOption(value) => {
-                if let Ok(update) = serde_json::from_value::<Vec<SessionConfigOption>>(value.clone()) {
-                    let mut s = self.session.write().await;
-                    s.apply_advertised_config_options(update);
-                    self.commit_session_changes(&mut s).await;
-                }
-            }
-            AgentStreamEvent::AcpContextUsage(value) => {
-                if let Ok(update) = serde_json::from_value::<UsageUpdate>(value.clone()) {
-                    let mut s = self.session.write().await;
-                    s.apply_context_usage(update);
-                }
-            }
-            _ => {}
-        }
-    }
-
     /// Drain pending domain events from the session aggregate and
     /// forward them to the persistence consumer via the mpsc channel.
-    async fn commit_session_changes(&self, session: &mut AcpSession) {
+    pub(super) async fn commit_session_changes(&self, session: &mut AcpSession) {
         for event in session.drain_events() {
             let _ = self.domain_event_tx.send(event).await;
         }
